reorganize fetcher files
