+++ frontend/src/services/auth/UserProvider.tsx
@@ -3,9 +3,9 @@
 // note that importing these individually allows us to mock them, otherwise mocks don't work :shrug:
 import debounce from "lodash/debounce";
 import noop from "lodash/noop";
-import { userFetcher } from "src/app/api/clientUserFetcher";
 import { UserSession } from "src/services/auth/types";
 import { UserContext } from "src/services/auth/useUser";
+import { userFetcher } from "src/services/fetch/fetchers/clientUserFetcher";
 import { isSessionExpired } from "src/utils/authUtil";
 
 import React, { useCallback, useEffect, useMemo, useState } from "react";
