+++ frontend/src/services/fetch/fetcherHelpers.ts
