+++ frontend/src/services/fetch/fetchers/userFetcher.ts
@@ -1,4 +1,4 @@
-import { postUserLogout } from "src/app/api/fetchers";
+import { postUserLogout } from "src/services/fetch/fetchers/fetchers";
 
 export const postLogout = async (token: string) => {
   const jwtAuthHeader = { "X-SGG-Token": token };
